refactor: 도메인 모델들 패키징
