+++ src/main/java/blackjack/domain/rule/BlackJackScoreRule.java
@@ -1,4 +1,6 @@
-package blackjack.domain;
+package blackjack.domain.rule;
+
+import blackjack.domain.card.Card;
 
 import java.util.List;
 import java.util.stream.Collectors;
